Moved user schema to config.  Added timestamps.
